A few style improvements. Major work on the styling inside question-dropdown.
Found an interesting way to fix really looong words from overflowing.
Had a lot of trouble finding a solution to getting the delete button at the top right of the dropdown.
Using jquery to hide at the beginning of page load worked but was jerky.
The current method of using a flexbox and changing display from 'none' to 'flex' when clicking the dropdown works but is also jerky.